Reordered things Changed docker-compose and other small things
