+++ api/auth/src/models/mongo-User.js
@@ -12,7 +12,7 @@ const userSchema = new mongoose.Schema({
     password: {
         type: String,
         require: [true, 'Please enter a password'],
-        minlength: [6, 'Minimum password length is 8 characters'],
+        minlength: [8, 'Minimum password length is 8 characters'],
     }
 });
