+++ config/README.md
@@ -1,18 +1,18 @@
 # config/
 
-Local configuration related to setup on local machine
+Local configuration related to setup on local machine.
 
 ## For the APIs
 
 ### For backend devs
 
-To install necessaries dependencies, run the followings
+To install necessaries dependencies, run the followings:
 
 auth-api
 
 ```
 npm i jest --save-dev
-npm i express bcrypt jsonwebtoken dotenv mongoose
+npm i express dotenv bcrypt jsonwebtoken mongoose mysql2 knex objection
 ```
 
 account-api
@@ -22,23 +22,23 @@ npm i jest --save-dev
 npm i express dotenv
 ```
 
-### Setting up Mysql on local machine
-
-Copy the .sql files to the docker container
-
-```
-docker cp file.sql container_name:/path/in/container/
-```
-
 ### Access the Mysql container
 
-Refer to the .env file for the password
+run:
+
+```
+docker exec -it CONTAINER_ID mysql -p
+```
+
+It will prompt you to enter a password.
+
+Refer to the .env file for the password.
 
 ## For the Fronted
 
 ### To run react native
 
-You must have expo installed
+You must have expo installed.
 
 To install:
 
@@ -54,9 +54,9 @@ npm start
 
 ### To view changes
 
-- Download expo go on your phone
-- If you do not have an account, please make one
-- Scan the QR code generated in the terminal
+- Download expo go on your phone.
+- If you do not have an account, please make one.
+- Scan the QR code generated in the terminal.
 
 If you have a Mac with Xcode installed, you may enable hot reload on your Mac.
